Change getData method - add map to return DriverDataInterface
